web/requirements-pip.txt
django==1.9.6
gunicorn==19.5.0
celery
flower
psycopg2==2.6.1
redis==2.10.5
channels
asgi_redis
asgiref
Pillow==3.2.0
djangorestframework==3.4
djoser==0.4.3
djangorestframework-jwt
django-jwt-auth
sphinx==1.4.1
dj-static==0.0.6
boto==2.40.0
django-storages-redux==1.3.2
awscli==1.10.28
python-memcached==1.57
django-debug-toolbar
django-extensions
django-audit-log
django-location-field
django-rest-swagger==0.3
django-filter
django-crispy-forms
django-cors-headers